Split amd-transpile into cohesive modules + address Copilot feedback
Refactor (no behavior change):
- Break the 1100-LOC monolithic `amd-transpile.ts` into 5 cohesive
  files under `amd-transpile/`:
  - `index.ts` — `transpileAmd` public entry
  - `identifier-rewriter.ts` — scope-aware AST walk class
  - `pattern-helpers.ts` — pure AST helpers shared by main pass + walker
  - `emit-helpers.ts` — pure string utilities for AMD wrapper text
  - `top-level-await.ts` — TLA detector

Copilot review fixes (PR #4573):
- package.json: `es-module-lexer` and `ts-node` were in
  `peerDependencies` (only used by bench scripts) — moved to
  `devDependencies` so consumers don't get spurious peer-dep warnings.
- `IdentifierRewriter`: declare `node.id.name` for `FunctionDeclaration`
  in its own function scope (not just `FunctionExpression`) so a
  recursive self-call inside `function f() { f(); }` isn't rewritten
  to the import accessor when `f` is also imported.
- `reExportStarSnippet`: emit `configurable: true` on the getter so
  duplicate `export * from 'foo'` re-exports of the same key don't
  throw on the second `Object.defineProperty`. Matches `defineGetter`
  / `defineLocalGetter`.
- `export default <expr>` comment: clarify that the deferred-setter
  pattern avoids a magic-string overlap and unifies the emit pattern,
  not TDZ. The expression still evaluates at its source position, so
  forward references to a later `const`/`class` are still a TDZ
  error — same as native ESM.
